Add booklists route
