+++ src/App.js
@@ -5,6 +5,7 @@ import {
 } from 'react-router-dom';
 import Books from './components/Books';
 import Categories from './components/Categories';
+import Booklists from './components/Booklists';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import Navbar from './components/Navbar';
 
@@ -14,6 +15,9 @@ function App() {
       <div className="container border shadow mt-5">
         <Navbar />
         <Switch>
+          <Route path="/booklists">
+            <Booklists />
+          </Route>
           <Route path="/books">
             <Books />
           </Route>
